+++ attribute-selectors/src/main/java/com/microej/example/mwt/attribute/selector/AbstractAttributeSelector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2022 MicroEJ Corp. All rights reserved.
+ * Copyright 2015-2024 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.mwt.attribute.selector;
@@ -26,7 +26,7 @@ public abstract class AbstractAttributeSelector implements Selector {
 	 * @param attribute
 	 *            the attribute to check.
 	 */
-	public AbstractAttributeSelector(String attribute) {
+	protected AbstractAttributeSelector(String attribute) {
 		this.attribute = attribute;
 	}