+++ immutable-stylesheet/src/main/java/com/microej/example/mwt/immutablestylesheet/stylesheet/CascadingStylesheet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015-2022 MicroEJ Corp. All rights reserved.
+ * Copyright 2015-2024 MicroEJ Corp. All rights reserved.
  * Use of this source code is governed by a BSD-style license that can be found with this software.
  */
 package com.microej.example.mwt.immutablestylesheet.stylesheet;
@@ -36,8 +36,8 @@ public class CascadingStylesheet implements Stylesheet {
 		private final Selector selector;
 		private final ImmutableStyle style;
 
-		// Never used because instances are immutable but necessary to compile.
-		private SelectorStyle(SelectorStyle selectorStyle) {
+		private SelectorStyle(SelectorStyle selectorStyle) { // NOSONAR Never used because instances are immutable but
+																// necessary to compile.
 			this.selector = selectorStyle.selector;
 			this.style = selectorStyle.style;
 		}
@@ -46,8 +46,8 @@ public class CascadingStylesheet implements Stylesheet {
 	private final ImmutableStyle defaultStyle;
 	private final SelectorStyle[] selectorsStyles;
 
-	// Never used because instance is immutable but necessary to compile.
-	private CascadingStylesheet(CascadingStylesheet stylesheet) {
+	private CascadingStylesheet(CascadingStylesheet stylesheet) { // NOSONAR Never used because instances are immutable
+																	// but necessary to compile.
 		this.defaultStyle = stylesheet.defaultStyle;
 		this.selectorsStyles = stylesheet.selectorsStyles;
 	}