+++ buffered-image-pool/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2025-03-27
+
+### Changed
+
+- Update Gradle plug-in to `1.1.0`.
+
+### Fixed
+
+- Update README and reset images heap to a value closer to the required one.
+
 ## [2.0.0] - 2024-03-05
 
 ### Changed
@@ -34,5 +44,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create an histogram widget that uses a screenshot on a buffered image to draw the background grid pattern.
 
 ---  
-_Copyright 2021-2024 MicroEJ Corp. All rights reserved._  
+_Copyright 2021-2025 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._