+++ mvc/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.1] - 2025-03-27
+
+### Changed
+
+- Make pointer tolerance dependent on the display size.
+- Update Gradle plug-in to `1.1.0`.
+
+### Fixed
+
+- Reset images heap to 0.
+
 ## [5.0.0] - 2024-03-05
 
 ### Changed
@@ -89,5 +100,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial revision.
 
 ---  
-_Copyright 2016-2024 MicroEJ Corp. All rights reserved._  
+_Copyright 2016-2025 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._