+++ popup/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2025-03-27
+
+### Changed
+
+- Update Gradle plug-in to `1.1.0`.
+
+### Fixed
+
+- Reset images heap to 0.
+
 ## [2.0.0] - 2024-03-05
 
 ### Changed
@@ -26,5 +36,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create an event dispatcher that sends events only to the popup while open.
 
 ---  
-_Copyright 2021-2024 MicroEJ Corp. All rights reserved._  
+_Copyright 2021-2025 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._