+++ transition/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2025-03-27
+
+### Changed
+
+- Update Gradle plug-in to `1.1.0`.
+
+### Fixed
+
+- Fix glitches that occur with predraw refresh strategy.
+- Update README and reset images heap to a value closer to the required one.
+
 ## [2.0.0] - 2024-03-05
 
 ### Changed
@@ -18,5 +29,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create a transition container that slides the incoming widget using a screenshot on a buffered image.
 
 ---  
-_Copyright 2023-2024 MicroEJ Corp. All rights reserved._  
+_Copyright 2023-2025 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._